Edit

+++ src/main/java/animals/herbivores/Herbivores.java
@@ -4,7 +4,8 @@ import animals.Animal;
 
 public abstract class Herbivores extends Animal {
 
-    public Herbivores(int maxSaturation, int saturation, int speed, int weight) {
+
+    public Herbivores(int maxSaturation, int saturation, int speed, double weight) {
         super(maxSaturation, saturation, speed, weight);
     }
 }
