+++ src/main/java/animals/predator/Predator.java
@@ -4,8 +4,8 @@ import animals.Animal;
 
 public abstract class Predator extends Animal {
 
-    public Predator(int maxSaturation, int saturation, int speed, int weight) {
+
+    public Predator(int maxSaturation, int saturation, int speed, double weight) {
         super(maxSaturation, saturation, speed, weight);
     }
-
 }
